django/15_项目/代码/day15/requirement.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.0.0
Pygments==2.5.2
pytz==2019.3
redis==3.4.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
